refactor: move game controller to separate folder
